refactor: extract highlightHtml to helper file and add disclaimer
